FIX: op_helper.c adjust/build cpu_{in|out}b()
Originally based on Mar 25, 2011 commit at "dirty hack to fix SIGSEGV
when doing IN, OUT and A is not zero."

https://github.com/legumbre/qemu-z80/commit/b14134a2

Update: QEmu v1's ioport.c API now accepts 16-bit port numbers
without the crash observed [in repo.or.cz's v0.10.50 fork], however
the Z80 'IN' and 'OUT' instructions should fill the eight high bits
of the port address with the accumulator value and then the called
functions should ignore these.

Leonardo Etcheverry's workaround for this was to leave out both of
these steps. We can avoid adjusting ioport.c (and the fprintf()s in
ioport-user.c) by retaining this change.