FIX: Avoid bogus uninitialised warning [1]
Pseudocode to reproduce:
1. declare variable and initialise under desired conditions
2. note that first reference under same conditions is acceptable
3. note that subsequent references under same conditions give warning

In one of our switch cases we ignore 'r2' completely, but in more than
one other code path we reference it again. gcc v6.3.0 seems capable of
working out the conditional use is under the same terms twice, but
loses track of its findings when it encounters a third (and any later)
reference(s). Avoid the warning that results.