WIP: i386-style z80_*() in/out helpers
Previously the passing of 16-bit port addresses to QEmu built-in
wrapper functions by IN and OUT instructions led to a segfault.

IN and OUT instructions as documented in Zilog's Z80 User Manual
provide 16 bits of data to the address bus, and includes the
accumulator register value in bits A8 to A15. The CPU "[only uses A0 to
A7] to select the I/O device at one of 256 possible ports", however.

Ensuring the wrapper function filters any ignored bits internally,
allows the original repo.or.cz implementation to be followed verbatim.