FIX: Avoid bogus uninitialised warning [2]
In these load instructions, we want to reference a delta value
fetched from RAM if either r1 or r2 represents a register that needs
it.

Accordingly, 'd' is assigned to under `COND_R1 || COND_R2` and then
referenced under COND_R1 or under COND_R2. These are not both true at
the same time, meaning we know we only assign once and then only
reference in the case an assignment was made earlier, however gcc
v6.3.0 struggles to see this. Avoid the warning that results.